requirements.txt
Flask==3.1.3
Flask-Login==0.6.3
Werkzeug==3.1.6
gspread==6.2.1
google-auth==2.49.1
google-auth-oauthlib==1.3.0
anthropic==0.86.0
requests==2.32.5
gunicorn==23.0.0
